requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
openai
python-dotenv
langchain
langchain-community
pdfplumber